Refactor evjs orchestration into ev and slim cli wrapper
