+++ packages/cli/README.md
@@ -1,6 +1,6 @@
 # @evjs/cli
 
-> CLI and configuration for the **evjs** fullstack framework.
+> Thin command-line wrapper for the **evjs** fullstack framework.
 
 ## Install
 
@@ -10,7 +10,7 @@ npm install -g @evjs/cli
 
 ## Convention over Configuration
 
-No configuration file is needed. `ev dev` and `ev build` work out of the box with sensible defaults:
+No configuration file is needed. `ev dev` and `ev build` delegate to `@evjs/ev` and inject the default utoopack adapter:
 
 - Entry: `./src/main.tsx`
 - HTML: `./index.html`
@@ -29,13 +29,13 @@ No configuration file is needed. `ev dev` and `ev build` work out of the box wit
 
 ### `ev dev`
 
-Uses bundler Node API directly (no temp config files):
+Uses the default bundler adapter directly (no temp config files):
 1. **dev server** (port 3000) — client bundle with HMR.
 2. **Node API Server** (port 3001) — auto-starts when server bundle is emitted, uses `node --watch`.
 
 ### `ev build`
 
-Runs webpack via Node API with `NODE_ENV=production`:
+Runs the production build through `@evjs/ev` with `NODE_ENV=production`:
 - `dist/client/` — optimized client assets with content hashes.
 - `dist/server/main.[hash].js` — server bundle (entry discovered via `dist/server/manifest.json`).
 
@@ -51,7 +51,6 @@ export default defineConfig({
   html: "./index.html",
   dev: { port: 3000 },
   server: {
-    endpoint: "/api/fn",
     dev: { port: 3001 },
   },
 });
@@ -80,13 +79,13 @@ my-app/
 ## Common Mistakes
 
 1. **Don't create `custom bundler config file`** — use `ev.config.ts` instead
-2. **Don't install webpack manually** — it's a dependency of `@evjs/cli`
+2. **Don't install bundler internals manually** — the default adapter is provided by `@evjs/cli`
 3. **Config file must be `ev.config.ts`** — not `evjs.config.ts`
 4. **Import `defineConfig` from `@evjs/ev`** — not from `@evjs/server`
 
 ## Bundled Dependencies
 
-Users do NOT need to install these — they're included in `@evjs/cli`:
-- `webpack`, `webpack-dev-server`
-- `html-webpack-plugin`, `swc-loader`, `@swc/core`
-- `@evjs/bundler-utoopack`, `@evjs/build-tools`
+Users do NOT need to install these — they're included through `@evjs/cli`:
+- `@evjs/bundler-utoopack`
+- `@evjs/build-tools`
+- the bundler's underlying compiler dependencies
