+++ packages/cli/tests/programmatic.test.ts
@@ -1,0 +1,46 @@
+import fs from "node:fs";
+import os from "node:os";
+import path from "node:path";
+import type { BundlerAdapter } from "@evjs/ev";
+import { describe, expect, it } from "vitest";
+import { build } from "../src/index.js";
+
+async function createProject() {
+  const cwd = await fs.promises.mkdtemp(path.join(os.tmpdir(), "evjs-cli-"));
+  await fs.promises.writeFile(
+    path.join(cwd, "index.html"),
+    '<div id="app"></div>',
+    "utf-8",
+  );
+  return cwd;
+}
+
+describe("programmatic API", () => {
+  it("forwards build calls through the framework API", async () => {
+    const cwd = await createProject();
+    const events: string[] = [];
+    const bundler: BundlerAdapter = {
+      name: "mock",
+      async build(_config, buildCwd) {
+        events.push(`build:${buildCwd}`);
+        const dist = path.join(buildCwd, "dist");
+        await fs.promises.mkdir(dist, { recursive: true });
+        await fs.promises.writeFile(
+          path.join(dist, "manifest.json"),
+          JSON.stringify({
+            version: 1,
+            assets: { js: [], css: [] },
+          }),
+          "utf-8",
+        );
+      },
+      async dev() {
+        events.push("dev");
+      },
+    };
+
+    await build({ server: false }, { cwd, bundler });
+
+    expect(events).toEqual([`build:${cwd}`]);
+  });
+});
