+++ packages/ev/tsconfig.json
@@ -11,6 +11,7 @@
     "stripInternal": true,
     "declarationMap": true,
     "sourceMap": true,
+    "types": ["node"],
     "verbatimModuleSyntax": true,
     "skipLibCheck": true,
     "strict": true,
